Make pathfinder reusable
